Preserve trailing spaces in poked files
Currently, using `xmlpoke` to modify a XML document discards all spaces that were present at the end of the source.
This is a nuisance when one wants all files to have a trailing newline character.

This change preserves all trailing spaces in a modified document.

IMHO `xmlpoke` is the right place for this feature, since it's about preserving spaces *outside* of the actual XML document. So one could argue that this is outside of the responsibility of `xmldom`.

The implementation is chosen such that, even if `xmldom` should implement this feature someday, we will not add more trailing spaces than in the original document.